+++ src/Infrastructure/Console/ContainerFactory.php
@@ -5,8 +5,13 @@ declare(strict_types=1);
 namespace Kinescope\Infrastructure\Console;
 
 use Kinescope\Core\ApiClientFactory;
-use Kinescope\Infrastructure\Console\Command\BrowseCommand;
-use Kinescope\Infrastructure\Console\Command\VideoInfoCommand;
+use Kinescope\Infrastructure\Console\Command\FolderListCommand;
+use Kinescope\Infrastructure\Console\Command\FolderShowCommand;
+use Kinescope\Infrastructure\Console\Command\ProjectListCommand;
+use Kinescope\Infrastructure\Console\Command\ProjectShowCommand;
+use Kinescope\Infrastructure\Console\Command\VideoAssetListCommand;
+use Kinescope\Infrastructure\Console\Command\VideoListCommand;
+use Kinescope\Infrastructure\Console\Command\VideoShowCommand;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -26,15 +31,20 @@ final class ContainerFactory
             ->setFactory([self::class, 'buildApiClientFactory'])
             ->addArgument(new Reference(LoggerInterface::class));
 
-        $container->register(VideoInfoCommand::class)
-            ->setPublic(true)
-            ->addArgument(new Reference(ApiClientFactory::class))
-            ->addArgument(new Reference(LoggerInterface::class));
-
-        $container->register(BrowseCommand::class)
-            ->setPublic(true)
-            ->addArgument(new Reference(ApiClientFactory::class))
-            ->addArgument(new Reference(LoggerInterface::class));
+        foreach ([
+            ProjectListCommand::class,
+            ProjectShowCommand::class,
+            FolderListCommand::class,
+            FolderShowCommand::class,
+            VideoListCommand::class,
+            VideoShowCommand::class,
+            VideoAssetListCommand::class,
+        ] as $commandClass) {
+            $container->register($commandClass)
+                ->setPublic(true)
+                ->addArgument(new Reference(ApiClientFactory::class))
+                ->addArgument(new Reference(LoggerInterface::class));
+        }
 
         $container->compile();